Refactoing header component. Adding header to login screen
